requirements.txt
streamlit
pandas
numpy
matplotlib>=3.8
google-generativeai>=0.7.0
python-dateutil
requests


